add threadFactoryWithPrefix and getThreadGroup
